Rename component -> components
