+++ src/components/Page/Base.js
@@ -2,7 +2,7 @@ import { NextSeo } from "next-seo";
 import PropTypes from "prop-types";
 import React from "react";
 
-import Footer from "@/pesayetu/component/Footer";
+import Footer from "@/pesayetu/components/Footer";
 import { footerArgs } from "@/pesayetu/config";
 /**
  * Base page that can be used to build all other pages.
