+++ src/pages/index.js
@@ -1,8 +1,8 @@
 import PropTypes from "prop-types";
 import React from "react";
 
-import ExploreOtherTools from "@/pesayetu/component/ExploreOtherTools";
-import Page from "@/pesayetu/component/Page";
+import ExploreOtherTools from "@/pesayetu/components/ExploreOtherTools";
+import Page from "@/pesayetu/components/Page";
 import { exploreTools } from "@/pesayetu/config";
 import getPostTypeStaticProps from "@/pesayetu/functions/postTypes/getPostTypeStaticProps";
